Broke up mapreduce module
